Refactor OrgTable
