+++ src/views/Orgs/Orgs.js
@@ -1,14 +1,15 @@
 import React, { Component, Fragment } from 'react'
-import { withStyles } from '@material-ui/core';
+import { withStyles, Paper, Button, DialogActions, ListItemText, ListItem, List, DialogContentText, DialogContent, DialogTitle, Dialog, ListItemIcon, IconButton } from '@material-ui/core';
 import projectStyles from 'assets/jss/views/projects';
 import CircularLoader from 'components/Loader/CircularLoader';
 import GridContainer from 'components/Grid/GridContainer';
 // import { getAllOrgs } from 'variables/dataOrgs';
 import OrgTable from 'components/Orgs/OrgTable';
 // import Toolbar from 'components/Toolbar/Toolbar'
-import { People, Business } from 'variables/icons';
+import { People, Business, PictureAsPdf, Delete, Edit, Star, StarBorder, Add } from 'variables/icons';
 import { filterItems, handleRequestSort } from 'variables/functions'
 import { deleteOrg } from 'variables/dataOrgs';
+import TableToolbar from 'components/Table/TableToolbar';
 
 class Orgs extends Component {
 	constructor(props) {
@@ -16,15 +17,15 @@ class Orgs extends Component {
 
 		this.state = {
 			orgs: [],
+			selected: [],
+			openDelete: false,
 			loading: true,
 			route: 1,
 			order: 'desc',
 			orderBy: 'name',
 			filters: {
 				keyword: '',
-				startDate: null,
-				endDate: null,
-				activeDateFilter: false
+				custom: []
 			}
 		}
 		props.setHeader('orgs.pageTitle', false, '', 'users')
@@ -33,6 +34,136 @@ class Orgs extends Component {
 		this.setState({ loading: true })
 		await this.props.reload()
 	}
+	options = () => {
+		const { t, accessLevel, isFav, orgs } = this.props
+		const { selected } = this.state
+		let org = orgs[orgs.findIndex(d => d.id === selected[0])]
+		let favObj = {
+			id: org.id,
+			name: org.name,
+			type: 'org',
+			path: `/management/org/${org.id}`
+		}
+		let isFavorite = isFav(favObj)
+		let allOptions = [
+			{ label: t('menus.edit'), func: this.handleEdit, single: true, icon: Edit },
+			{ label: isFavorite ? t('menus.favorites.remove') : t('menus.favorites.add'), icon: isFavorite ? Star : StarBorder, func: isFavorite ? () => this.removeFromFav(favObj) : () => this.addToFav(favObj) },
+			{ label: t('menus.exportPDF'), func: () => { }, icon: PictureAsPdf },
+			{ label: t('menus.delete'), func: this.handleOpenDeleteDialog, icon: Delete }
+		]
+		if (accessLevel.apiorg.edit)
+			return allOptions
+		else return [
+			{ label: isFavorite ? t('menus.favorites.remove') : t('menus.favorites.add'), icon: isFavorite ? Star : StarBorder, func: isFavorite ? () => this.removeFromFav(favObj) : () => this.addToFav(favObj) },
+			{ label: t('menus.exportPDF'), func: () => { }, icon: PictureAsPdf }
+		]
+	}
+	ftUsers = () => {
+		const { t } = this.props
+		return [
+			{ key: 'firstName', name: t('users.fields.firstName'), type: 'string' },
+			{ key: 'lastName', name: t('users.fields.lastName'), type: 'string' },
+			{ key: 'email', name: t('users.fields.email'), type: 'string' },
+			{ key: 'org.name', name: t('orgs.fields.name'), type: 'string' },
+			// { key: 'devices[0].start', name: t('collections.fields.activeDeviceStartDate'), type: 'date' },
+			// { key: 'created', name: t('collections.fields.created'), type: 'date' },
+			// { key: 'activeDeviceStats.state', name: t('devices.fields.status'), type: 'dropDown', options: this.dLiveStatus() }
+			// { key: 'address', name: t('devices.fields.address'), type: 'string' },
+			// { key: 'locationType', name: t('devices.fields.locType'), type: 'dropDown', options: this.dLocationPlace() },
+			// { key: 'lat', name: t('calibration.stepheader.calibration'), type: 'diff', options: { dropdown: this.dCalibrated(), values: { false: [0] } } },
+		]
+	}
+	addFilter = (f) => {
+		let cFilters = this.state.filters.custom
+		let id = cFilters.length
+		cFilters.push({ ...f, id: id })
+		this.setState({
+			filters: {
+				...this.state.filters,
+				custom: cFilters
+			}
+		})
+		return id
+	}
+	removeFilter = (fId) => {
+		let cFilters = this.state.filters.custom
+		cFilters = cFilters.reduce((newFilters, f) => {
+			if (f.id !== fId) {
+				newFilters.push(f)
+			}
+			return newFilters
+		}, [])
+		this.setState({
+			filters: {
+				...this.state.filters,
+				custom: cFilters
+			}
+		})
+	}
+	handleCheckboxClick = (event, id) => {
+		event.stopPropagation()
+		const { selected } = this.state;
+		const selectedIndex = selected.indexOf(id)
+		let newSelected = [];
+
+		if (selectedIndex === -1) {
+			newSelected = newSelected.concat(selected, id);
+		} else if (selectedIndex === 0) {
+			newSelected = newSelected.concat(selected.slice(1))
+		} else if (selectedIndex === selected.length - 1) {
+			newSelected = newSelected.concat(selected.slice(0, -1))
+		} else if (selectedIndex > 0) {
+			newSelected = newSelected.concat(
+				selected.slice(0, selectedIndex),
+				selected.slice(selectedIndex + 1),
+			);
+		}
+
+		this.setState({ selected: newSelected })
+	}
+	handleOpenDeleteDialog = () => {
+		this.setState({ openDelete: true, anchorElMenu: null })
+	}
+
+	handleEdit = () => {
+		this.props.history.push(`/management/org/${this.state.selected[0]}/edit`)
+	}
+	addToFav = (favObj) => {
+		this.props.addToFav(favObj)
+		this.setState({ anchorElMenu: null })
+	}
+	removeFromFav = (favObj) => {
+		this.props.removeFromFav(favObj)
+		this.setState({ anchorElMenu: null })
+	}
+	handleCloseDeleteDialog = () => {
+		this.setState({ openDelete: false })
+	}
+	handleDeleteOrg = async () => {
+		await this.handleDeleteOrgs()
+		this.setState({
+			selected: [],
+			anchorElMenu: null,
+			openDelete: false
+		})
+	}
+	handleSelectAllClick = (event, checked) => {
+		if (checked) {
+			this.setState({ selected: this.props.orgs.map(n => n.id) })
+			return;
+		}
+		this.setState({ selected: [] })
+	}
+	handleToolbarMenuOpen = e => {
+		e.stopPropagation()
+		this.setState({ anchorElMenu: e.currentTarget })
+	}
+
+	handleToolbarMenuClose = e => {
+		e.stopPropagation();
+		this.setState({ anchorElMenu: null })
+	}
+
 	orgsHeader = () => {
 		const { t } = this.props
 		return [
@@ -110,6 +241,9 @@ class Orgs extends Component {
 		{ id: 0, title: this.props.t('users.tabs.users'), label: <People />, url: `/management/users` },
 		{ id: 1, title: this.props.t('users.tabs.orgs'), label: <Business />, url: `/management/orgs` },
 	]
+	
+	addNewOrg = () => { this.props.history.push('/management/orgs/new') }
+
 	snackBarMessages = (msg) => {
 		const { s } = this.props
 		switch (msg) {
@@ -126,30 +260,86 @@ class Orgs extends Component {
 	handleTabsChange = (e, value) => {
 		this.setState({ route: value })
 	}
-	handleDeleteOrgs = async (selected) => {
+	handleDeleteOrgs = async () => {
+		const { selected } = this.state
 		await selected.forEach(async u => {
 			await deleteOrg(u)
 		})
 		await this.props.reload()
 		this.snackBarMessages(1)
 	}
+	renderConfirmDelete = () => {
+		const { openDelete, selected } = this.state
+		const { orgs, t, classes } = this.props
+		return <Dialog
+			open={openDelete}
+			onClose={this.handleCloseDeleteDialog}
+			aria-labelledby='alert-dialog-title'
+			aria-describedby='alert-dialog-description'
+		>
+			<DialogTitle id='alert-dialog-title'>{t('dialogs.delete.title.orgs')}</DialogTitle>
+			<DialogContent>
+				<DialogContentText id='alert-dialog-description'>
+					{t('dialogs.delete.message.orgs')}:
+				</DialogContentText>
+				<List>
+					{selected.map(s => <ListItem classes={{ root: classes.deleteListItem }} key={s}><ListItemIcon><div>&bull;</div></ListItemIcon>
+						<ListItemText primary={orgs[orgs.findIndex(d => d.id === s)].name} /></ListItem>)}
+				</List>
+			</DialogContent>
+			<DialogActions>
+				<Button onClick={this.handleCloseDeleteDialog} color='primary'>
+					{t('actions.no')}
+				</Button>
+				<Button onClick={this.handleDeleteOrg} color='primary' autoFocus>
+					{t('actions.yes')}
+				</Button>
+			</DialogActions>
+		</Dialog>
+	}
+	renderTableToolBarContent = () => {
+		const { accessLevel } = this.props
+		let access = accessLevel.apiorg ? accessLevel.apiorg.edit ? true : false : false
+		return <Fragment>
+			{access ? <IconButton aria-label='Add new organisation' onClick={this.addNewOrg}>
+				<Add />
+			</IconButton> : null
+			}
+		</Fragment>
+	}
+
 	renderOrgs = () => {
-		const { t } = this.props
-		const { loading, order, orderBy, orgs, filters } = this.state
+		const { t, classes } = this.props
+		const { loading, order, orderBy, orgs, filters, selected } = this.state
 		return <GridContainer justify={'center'}>
-			{loading ? <CircularLoader /> : <OrgTable
-				data={orgs}
-				tableHead={this.orgsHeader()}
-				handleFilterEndDate={this.handleFilterEndDate}
-				handleFilterKeyword={this.handleFilterKeyword}
-				handleFilterStartDate={this.handleFilterStartDate}
-				handleRequestSort={this.handleRequestSort}
-				handleDeleteOrgs={this.handleDeleteOrgs}
-				orderBy={orderBy}
-				order={order}
-				filters={filters}
-				t={t}
-			/>}
+			{loading ? <CircularLoader /> :
+				<Paper className={classes.root}>
+					{this.renderConfirmDelete()}
+					<TableToolbar //	./TableToolbar.js
+						anchorElMenu={this.state.anchorElMenu}
+						handleToolbarMenuClose={this.handleToolbarMenuClose}
+						handleToolbarMenuOpen={this.handleToolbarMenuOpen}
+						numSelected={selected.length}
+						options={this.options}
+						t={t}
+						content={this.renderTableToolBarContent()}
+					/>
+					<OrgTable
+						data={orgs}
+						tableHead={this.orgsHeader()}
+						handleFilterEndDate={this.handleFilterEndDate}
+						handleFilterKeyword={this.handleFilterKeyword}
+						handleFilterStartDate={this.handleFilterStartDate}
+						handleRequestSort={this.handleRequestSort}
+						handleDeleteOrgs={this.handleDeleteOrgs}
+						handleCheckboxClick={this.handleCheckboxClick}
+						handleSelectAllClick={this.handleSelectAllClick}
+						orderBy={orderBy}
+						selected={selected}
+						order={order}
+						filters={filters}
+						t={t}
+					/></Paper>}
 		</GridContainer>
 	}
 	render() {
